Add filtering by kind when graphing (-kinds)
The filtering is done with a function that returns true in case the api.Object
should be graphed and false otherwise. At the moment this function is just
being used for graphing by kind but in the future it can be used for other
things like for example filter by name.

The argument to the command was moved from -api to -kinds.