Make store.Store to be store.Graph
Originally store.Store was embedding store.Graph which itself was
graph.Graph (gonum.Graph). This was polluting the store interfaces and
frankly wasn't ideal as it required implementation of some functions
which were not necessary for the program functionality. This commit
cleans this up. This has also a benefit of having application specific
node IDs as opposed to the number id required by gonum.Graph which are
often hard to extract from graph databases and frankly they really
should not be!

We've also cleaned up memory store tests and added a bit more coverage.